(platform) Change the variable type for timer duration to long
This change modifies the timer duration type from an unsigned integer to
a signed long. This is to fix an issue where overflow could occur when
converting the expiration time value received in SIP messages (i.e.
REGISTER, INVITE, SUBSCRIBE, PUBLISH) from the network to a millisecond
value, particularly when the expiration time is very large.

Bug: 465295963
Flag: NONE pre-production
Test: atest ImsStackJavaTests, ImsStackNativeTests
Test: manual (verify the timer is set correctly when receiving a very
large expiration time value)

+++ java/src/com/android/imsstack/system/SystemInterface.java
@@ -491,7 +491,7 @@ public class SystemInterface implements JniSystemListener {
 
     private void handleSystemCallForTimer(int method, Parcel in, Parcel out) {
         if (method == SystemConstants.SET_TIMER) {
-            int duration = in.readInt();
+            long duration = in.readLong();
             long tid = in.readLong();
             out.writeInt(mDefaultSystemCall.startTimer(tid, duration) ? 1 : 0);
         } else if (method == SystemConstants.KILL_TIMER) {
